Refactor: treat multi-definition of output_sinlge_R
