+++ source/module_io/cal_r_overlap_R.h
@@ -17,7 +17,7 @@
 #include "module_base/vector3.h"
 #include "module_base/abfs-vector3_order.h"
 #include "module_base/ylm.h"
-#include "write_HS_sparse.h"
+#include "single_R_io.h"
 
 
 // output r_R matrix, added by Jingan
